requirements-test.txt
Pillow==3.2.0
coverage==4.0.3
django-simple-captcha==0.5.1
djangorestframework==3.3.3
pluggy==0.3.1
py==1.4.31
pytest==2.9.1
pytest-cov==2.2.1
pytest-django==2.9.1
six==1.10.0
tox==2.3.1
virtualenv==15.0.1
